requirements.txt
anthropic>=0.25.0
requests>=2.31.0
scikit-optimize>=0.9.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
tenacity>=8.2.0
matplotlib>=3.7.0
rich>=13.0.0
# For real Vapi call validation server
fastapi>=0.110.0
uvicorn>=0.29.0
